feat: DeleteInterfaceTrait is added
